Caesar Cipher
A script you can use to encode or decode a message between friends, family, or colleagues. Simply run the script, choose between encode or decode, choose a message to encode or decode, choose the amount you'd like to shift each letter, then you get your result. Characters are not supported yet and will be appended to the end of the result.