Silent auction
Script you can use with your friends to perform an auction. All you do is state your name and your bid, then the program will ask if anyone else is bidding. If not, it'll return you as the winner. For 2 or more, it returns the person with the highest bid. #day9